cross: update build.ocp and 00_config.ocp for the cross-compilation
We updated the build.ocp and 00_config.ocp for the cross-compilation,
and added a rule in the Makefile to generate a flow executable from
Linux for Windows (make build-flow-with-mingw)

We added a README.cross which explain how to use the cross-compiler.

We also replaced all `Sys.win32` by `Sys_utils.win32` and renamed
"Windows.h" by "windows.h".